refactor: mv ExtraComponentReference-Label-definitions -> ocm.gardener
ExtraComponentReference-Label is (in spite of chosen label-name)
Gardener-(CICD-)specific. Hence, mv it to ocm.gardener module, as
another prepration towards dropping `cnudie`, but also making code from
Concourse-Pipeline-Template re-usable (in this case: resolving of
component-references).
