+++ ctt/process_dependencies.py
@@ -24,6 +24,7 @@ import cnudie.iter
 import cnudie.retrieve
 import container.util
 import ocm
+import ocm.gardener
 import oci
 import oci.client
 import oci.model as om
@@ -269,11 +270,11 @@ def determine_changed_components(
     if not (
         reftype_filter and reftype_filter(cnudie.iter.NodeReferenceType.EXTRA_COMPONENT_REFS_LABEL)
     ) and (
-        extra_crefs_label := component.find_label(cnudie.iter.ExtraComponentReferencesLabel.name)
+        extra_crefs_label := component.find_label(ocm.gardener.ExtraComponentReferencesLabel.name)
     ):
         for extra_cref_raw in extra_crefs_label.value:
             extra_cref = dacite.from_dict(
-                data_class=cnudie.iter.ExtraComponentReference,
+                data_class=ocm.gardener.ExtraComponentReference,
                 data=extra_cref_raw,
             )
             extra_cref_id = extra_cref.component_reference
@@ -478,7 +479,7 @@ def process_images(
     executor = concurrent.futures.ThreadPoolExecutor(max_workers=16)
 
     reftype_filter = None
-    if remove_label and remove_label(cnudie.iter.ExtraComponentReferencesLabel.name):
+    if remove_label and remove_label(ocm.gardener.ExtraComponentReferencesLabel.name):
         def filter_extra_component_refs(reftype: cnudie.iter.NodeReferenceType) -> bool:
             return reftype is cnudie.iter.NodeReferenceType.EXTRA_COMPONENT_REFS_LABEL
